# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "AI Chat API",
    "version": "1.0.0",
    "description": "API documentation for the AI Chat backend. All endpoints require Clerk authentication."
  },
  "servers": [
    { "url": "http://localhost:3000/api" }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
      "Message": {
        "type": "object",
        "properties": {
          "_id": { "type": "string" },
          "conversationId": { "type": "string" },
          "sender": { "type": "string", "enum": ["user", "ai"] },
          "text": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "Conversation": {
        "type": "object",
        "properties": {
          "_id": { "type": "string" },
          "clerkUserId": { "type": "string" },
          "title": { "type": "string" },
          "messages": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Message" }
          },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      }
    }
  },
  "security": [{ "bearerAuth": [] }],
  "paths": {
    "/chat/": {
      "get": {
        "summary": "Get all conversations for the authenticated user",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "List of conversations",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Conversation" }
                }
              }
            }
          },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/chat/{id}": {
      "get": {
        "summary": "Get all messages for a specific conversation",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "List of messages",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Message" }
                }
              }
            }
          },
          "401": { "description": "Unauthorized" },
          "404": { "description": "Conversation not found" }
        }
      }
    },
    "/chat/send": {
      "post": {
        "summary": "Send a new message and get AI response",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "text": { "type": "string" },
                  "conversationId": { "type": "string" }
                },
                "required": ["text"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User and AI messages returned",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "userMessage": { "$ref": "#/components/schemas/Message" },
                    "aiMessage": { "$ref": "#/components/schemas/Message" }
                  }
                }
              }
            }
          },
          "400": { "description": "Validation error" },
          "401": { "description": "Unauthorized" },
          "404": { "description": "Conversation not found" }
        }
      }
    }
  }
} 